+++ CHANGELOG.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## v1.4.6 - 2024-06-XX
+## v1.4.7 - 2024-06-05
+
+### Added
+
+- more types
+
+### Fixed
+
+- publish of events that exceeds the DB driver limit are chunked now
+
+## v1.4.6 - 2024-05-29
 
 ### Added